% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateG.R
\name{estimateG}
\alias{estimateG}
\title{estimateG}
\usage{
estimateG(validFold, folds, L0, L1, A0, A1, abar, SL.g, glm.g, stratify,
  return.models, SL.g.options, verbose, tolg, ...)
}
\arguments{
\item{L0}{A \code{data.frame} featuring covariates measured at baseline.}

\item{L1}{A \code{data.frame} featuring time-varying covariates measured at 
the first timepoint.}

\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}

\item{SL.g}{A \code{vector} or \code{list} specifying the SuperLearner library
to be used to estimate the conditional probability of treatment at each time point.  See \code{SuperLearner}
package for details.}

\item{glm.g}{A \code{character} specifying the right-hand side of the \code{glm} 
formula used to estimate the conditional probability of treatment at each time point. 
Only used if \code{SL.g = NULL}.}

\item{stratify}{A \code{boolean} indicating whether to pool across treatment
nodes or to estimate outcome regression separately in each category.}

\item{return.models}{A \code{boolean} indicating whether the models for g00 should be 
returned with the output.}

\item{SL.g.options}{A \code{list} of additional arguments passed to \code{SuperLearner} for condtional treatment 
probability fits.}

\item{tolg}{A \code{numeric} indicating the truncation level for conditional treatment probabilities.}

\item{...}{Other arguments (currently passed to \code{SuperLearner}).}
}
\value{
Returns a list with \code{g0n}, \code{g1n}, and the estimated model objects if
\code{return.models = TRUE}
}
\description{
This function computes the conditional treatment probabilities at both timepoints.
}
\examples{
TO DO : add examples
}
