% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQ1.ltmle.R
\name{estimateQ1.ltmle}
\alias{estimateQ1.ltmle}
\title{estimateQ1.ltmle}
\usage{
estimateQ1.ltmle(L2, Q2n, L0, A0, A1, folds, validFold, abar, SL.Q,
  SL.Q.options, glm.Q, glm.Q.options, return.models, verbose, stratify, ...)
}
\arguments{
\item{L2}{A \code{vector} outcome of interest}

\item{L0}{A \code{data.frame} featuring covariates measured at baseline.}

\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{folds}{Vector of cross-validation folds}

\item{validFold}{Which fold is validation fold}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}

\item{SL.Q}{A \code{vector} or \code{list} specifying the SuperLearner library
to be used to estimate the conditional probability of treatment at each time point.  
See \code{SuperLearner} package for details.}

\item{SL.Q.options}{A \code{list} of additional arguments passed to \code{SuperLearner} 
for outcome regression fits.}

\item{glm.Q}{A \code{character} specifying the right-hand side of the \code{glm} 
formula used to estimate the conditional probability of treatment at each time point. 
Only used if \code{SL.Q = NULL}.}

\item{glm.Q.options}{A \code{list} of additional arguments passed to \code{glm} for the
outcome regression fits. Typically, the \code{family} argument.}

\item{return.models}{A \code{boolean} indicating whether the models for g00 should be 
returned with the output.}

\item{stratify}{A \code{boolean} indicating whether to pool across treatment
nodes or to estimate outcome regression separately in each category.}

\item{...}{Other arguments (not currently used).}

\item{L1}{A \code{data.frame} featuring time-varying covariates measured at 
the first timepoint.}
}
\value{
Returns a list with \code{Q2n}, \code{Q1n}, and the estimated model objects if
\code{return.models = TRUE}
}
\description{
This function computes the conditional treatment probabilities at time 1
}
\examples{
TO DO : add examples
}
