% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQr.R
\name{estimateQr}
\alias{estimateQr}
\title{estimateQr}
\usage{
estimateQr(rQ1, rQ2, g0n, g1n, A0, A1, SL.Qr, abar, returnModels, ...)
}
\arguments{
\item{rQ1}{The "residual" for the first reduced dimension regression (on g0n), 
equal to sum of the EIF at time 1 and 2.}

\item{rQ2}{The "residual" for the second reduced dimension regression (on g1n),
equal to the EIF at time 2.}

\item{g0n}{A \code{vector} of estimates of g_{0,0}.}

\item{g1n}{A \code{vector} of estimates of g_{1,0}.}

\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{SL.Qr}{A \code{vector} or \code{list} specifying the SuperLearner library
to be used to estimate the reduced-dimension regression to protect against misspecification of the
outcome regressions.  See \code{SuperLearner} package for details.}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}

\item{returnModels}{A \code{boolean} indicating whether the models for Qr0 should be 
returned with the output.}
}
\value{
A list with elements Q2nr, Q1nr, Q2mod, and Q1mod, corresponding to the
predicted values of the reduced dimension regressions and the models used to fit
them, respectively.
}
\description{
A function used to estimate the reduced dimension regressions for Q
}

