% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateEIF.R
\name{evaluateEIF}
\alias{evaluateEIF}
\title{evaluateEIF}
\usage{
evaluateEIF(A0, A1, L2, Q2n, Q1n, g1n, g0n, abar, ...)
}
\arguments{
\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{L2}{A \code{vector} outcome of interest.}

\item{Q2n}{A \code{vector} of estimates of Q_{2,0}}

\item{Q1n}{A \code{vector} of estimates of Q_{1,0}}

\item{g1n}{A \code{vector} of estimates of g_{1,0}}

\item{g0n}{A \code{vector} of estimates of g_{0,0}}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}

\item{...}{Additional arguments (not currently used)}
}
\value{
A list with named entries corresponding to the different pieces of the 
EIF at the nuisance parameter estimates evaluated at the observations.
}
\description{
Function that returns estimated efficient influence function contributions.
}

