% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateIF.R
\name{evaluateIF}
\alias{evaluateIF}
\title{evaluateIF}
\usage{
evaluateIF(A0, A1, L2, Q2n, Q1n, g1n, g0n, Q2nr.obsa, Q1nr, g0nr, g1nr, h0nr,
  h1nr, hbarnr, abar)
}
\arguments{
\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{L2}{A \code{vector} outcome of interest.}

\item{Q2n}{A \code{vector} of estimates of Q_{2,0}}

\item{Q1n}{A \code{vector} of estimates of Q_{1,0}}

\item{g1n}{A \code{vector} of estimates of g_{1,0}}

\item{g0n}{A \code{vector} of estimates of g_{0,0}}

\item{Q2nr.obsa}{A \code{vector} of estimates of Q_{2,0}^r at the observed values of A0.}

\item{Q1nr}{A \code{vector} of estimates of Q_{1,0}^r}

\item{g0nr}{A \code{vector} of estimates of g_{0,0}^r}

\item{g1nr}{A \code{vector} of estimates of g_{0,0}^r}

\item{h0nr}{A \code{vector} of estimates of h_{0,0}^r}

\item{h1nr}{A \code{vector} of estimates of h_{1,0}^r}

\item{hbarnr}{A \code{vector} of estimates of h^r, the iterated reduced
dimension regression.}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}
}
\value{
A list with named entries corresponding to the different pieces of the 
EIF at the nuisance parameter estimates evaluated at the observations.
}
\description{
Function that returns estimated influence function contributions.
}

