% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_replace.R
\name{list_replace}
\alias{list_replace}
\title{Helper function to replace values in a list}
\usage{
list_replace(x, list, values)
}
\arguments{
\item{x}{a list}

\item{list}{indices of list}

\item{values}{what values to replace with}
}
\value{
x with list values replaced
}
\description{
Helper function to replace values in a list
}
