% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{makeData.old}
\alias{makeData.old}
\title{makeData.old}
\usage{
makeData.old(n, b = 0.25, b0 = 1, ba = -0.25, setA = NULL)
}
\arguments{
\item{b}{Interaction coefficient for L01 and L11}

\item{ba}{The effect of each treatment on the outcome. See \code{makeData}
to see how it is used.}

\item{Y}{The outcome}

\item{X}{The predictors}

\item{newX}{The prediction frame}

\item{family}{Not used, just for compatibility}

\item{obsWeights}{Not used, just for compatibility}

\item{r}{The exponent of the rate at which error goes to zero as function of n.}
}
\value{
A \code{list} with L0, A0, L1, A1, L2 to input into \code{drinf.tmle}.
}
\description{
A function to simulate data for use in the simulation study. The function
can be used to generate observed data or data from an intervened SCM; the latter
is achieved via the \code{setA} option. The data are simulated as follows: 
\itemize{
\item L0.1 ~ Uniform(-4,4)
\item L0.2 ~ Bernoulli(1/2)
\item A0 ~ Bernoulli(expit(b0 + b*L0.1 - 2b*L0.2*L0.1))
\item L1.1 ~ Uniform(-4,4)
\item L1.2 ~ Bernoulli(1/2)
\item A1 ~ Bernoulli(expit(2*b0 + b*L0.1 - 2*b*L0.2*L0.1 + b/2*L1.1 - b*L1.2*L1.1))
\item L2 ~ Bernoulli(expit(ba*A0 + ba*A1 + b*L0.1 - 2*b*L0.2*L0.1 + b*L1.1 - 2*b*L1.2*L1.1))
}
}

