% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.CC_LS_mod.R
\name{method.CC_LS_mod}
\alias{method.CC_LS_mod}
\title{A modification of method.CC_LS_mod to 
not fail with duplicated predictions}
\usage{
method.CC_LS_mod()
}
\description{
A modification of method.CC_LS_mod to 
not fail with duplicated predictions
}
