% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.CC_LS_mod.R
\name{method.CC_nloglik_mod}
\alias{method.CC_nloglik_mod}
\title{Modification of convex combo neg log like that doesn't
fail with duplicated predictors}
\usage{
method.CC_nloglik_mod()
}
\description{
Modification of convex combo neg log like that doesn't
fail with duplicated predictors
}
