% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redReg.R
\name{redReg}
\alias{redReg}
\title{redReg}
\usage{
redReg(L2, A0, A1, Qn, gn, abar, verbose, return.models, SL.Qr, SL.gr, tolg,
  ...)
}
\value{
A list with Qnr and gnr objects. See \code{estimateQr} and \code{estimategr}
for details.
}
\description{
Function that computes the reduced dimension regressions needed for the 
extra targeting steps of the tmle.
}

