% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residG.R
\name{residG}
\alias{residG}
\title{residG}
\usage{
residG(A0, A1, g0n, g1n, abar, ...)
}
\arguments{
\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{g0n}{A \code{vector} of estimates of g_{0,0}}

\item{g1n}{A \code{vector} of estimates of g_{1,0}}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}

\item{...}{Other arguments (not currently used)}
}
\value{
A list with elements rg1 and rg2 that are the outcomes in the
reduced dimension regressions on Q1n and Q2n, respectively.
}
\description{
This function computes the residuals that are used as outcomes in the 
reduced dimension regressions for g.
}
