% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residQ.R
\name{residQ}
\alias{residQ}
\title{residQ}
\usage{
residQ(L2, A0, A1, Q2n, Q1n, g0n, g1n, abar, ...)
}
\arguments{
\item{L2}{A \code{vector} outcome of interest}

\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{Q2n}{A \code{vector} of estimates of Q_{2,0}}

\item{Q1n}{A \code{vector} of estimates of Q_{1,0}}

\item{g0n}{A \code{vector} of estimates of g_{0,0}}

\item{g1n}{A \code{vector} of estimates of g_{1,0}}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}
}
\value{
A list with elements rQ1 and rQ2 that are the outcomes in the
reduced dimension regressions on g0n and g1n, respectively.
}
\description{
This function computes the residuals that are used as outcomes in the 
reduced dimension regressions for Q.
}
