% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetQ1.ltmle.R
\name{targetQ1.ltmle}
\alias{targetQ1.ltmle}
\title{targetQ1.ltmle}
\usage{
targetQ1.ltmle(A0, A1, L2, Qn, gn, abar, tolg, tolQ, return.models,
  tol.coef = 10, ...)
}
\arguments{
\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{L2}{A \code{vector} outcome of interest.}

\item{Qn}{A \code{list} of current estimates of Q2n and Q1n}

\item{gn}{A \code{list} of current estimates of g1n and g0n}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}

\item{tolg}{A \code{numeric} indicating the truncation level for conditional treatment probabilities.}

\item{tolQ}{A \code{numeric}}

\item{return.models}{A \code{boolean} indicating whether the fluctuation model should be 
returned with the output.}

\item{tol.coef}{A \code{numeric} indicating the coefficient above which the minimization along the
submodel using \code{glm} is deemed to be unreasonable. In these cases \code{optim} is used 
instead to perform the fluctuation along the same submodel.}
}
\value{
A list with named entries corresponding to the estimators of the 
fluctuated nuisance parameters evaluated at the observed data values. If 
\code{return.models = TRUE} output also includes the fitted fluctuation model.
}
\description{
Function that targets Q2n for ltmle.
}

