% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetQg.ltmle.R
\name{targetQg.ltmle}
\alias{targetQg.ltmle}
\title{targetQg.ltmle}
\usage{
targetQg.ltmle(A0, A1, L2, Q2n, Q1n, g1n, g0n, abar, tolQ, return.models, ...)
}
\arguments{
\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{L2}{A \code{vector} outcome of interest.}

\item{Q2n}{A \code{vector} of estimates of Q_{2,0}}

\item{Q1n}{A \code{vector} of estimates of Q_{1,0}}

\item{g1n}{A \code{vector} of estimates of g_{1,0}}

\item{g0n}{A \code{vector} of estimates of g_{0,0}}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}

\item{tolQ}{A \code{numeric}}

\item{return.models}{A \code{boolean} indicating whether the fluctuation model should be 
returned with the output.}
}
\value{
A list with named entries corresponding to the estimators of the 
fluctuated nuisance parameters evaluated at the observed data values. If 
\code{return.models = TRUE} output also includes the fitted fluctuation model.
}
\description{
Function that targets estimates of Q to solve the EIF. Fits a single logistic 
fluctuation regression of c(Q2n, L2) on relevant offsets and so-called
clever covariates.
}
