% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetQg.R
\name{targetQg}
\alias{targetQg}
\title{targetQg}
\usage{
targetQg(A0, A1, L2, Qn, gn, Qnr.gnr, abar, tolg, tolQ, return.models,
  offset.model = TRUE, allatonce = FALSE, ...)
}
\arguments{
\item{A0}{A \code{vector} treatment delivered at baseline.}

\item{A1}{A \code{vector} treatment deliver after \code{L1} is measured.}

\item{L2}{A \code{vector} outcome of interest.}

\item{abar}{A \code{vector} of length 2 indicating the treatment assignment 
that is of interest.}

\item{tolg}{A \code{numeric} indicating the truncation level for conditional treatment probabilities.}

\item{tolQ}{A \code{numeric}}

\item{return.models}{A \code{boolean} indicating whether the fluctuation model should be 
returned with the output.}

\item{offset.model}{A \code{boolean} indicating whether to fit a fluctuation model
with an offset term (if \code{TRUE}) or intercept-only with weights (if \code{FALSE})}

\item{allatonce}{Do all the targeting with a single fit; if \code{FALSE}, then target g, followed by Q}

\item{Q2n}{A \code{vector} of estimates of Q_{2,0}}

\item{Q1n}{A \code{vector} of estimates of Q_{1,0}}

\item{g1n}{A \code{vector} of estimates of g_{1,0}}

\item{g0n}{A \code{vector} of estimates of g_{0,0}}

\item{Q2nr.obsa}{A \code{vector} of estimates of Q_{2,0}^r evaluated at observed A0 value}

\item{Q2nr.seta}{A \code{vector} of estimates of Q_{2,0}^r evaluated at A0 = abar[1]}

\item{Q1nr}{A \code{vector} of estimates of Q_{1,0}^r}

\item{g0nr}{A \code{vector} of estimates of g_{0,0}^r}

\item{g1nr}{A \code{vector} of estimates of g_{0,0}^r}

\item{h0nr}{A \code{vector} of estimates of h_{0,0}^r}

\item{h1nr}{A \code{vector} of estimates of h_{1,0}^r}

\item{hbarnr}{A \code{vector} of estimates of h^r, the iterated reduced
dimension regression.}
}
\value{
A list with named entries corresponding to the estimators of the 
fluctuated nuisance parameters evaluated at the observed data values. If 
\code{return.models = TRUE} output also includes the fitted fluctuation model.
}
\description{
Function that targets estimates of Q and g to solve the EIF + remainder 
approximations that result from misspecification. Fits a single logistic 
fluctuation regression of c(A0, A1, Q2n, L2) on relevant offsets and so-called
clever covariates.
}

