% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthG.R
\name{truthG}
\alias{truthG}
\title{truthG}
\usage{
truthG(Y, X, newX, family, obsWeights, b = 0.5, b0 = 1, r = -1/4,
  cons = 2, ...)
}
\arguments{
\item{Y}{The outcome}

\item{X}{The predictors}

\item{newX}{The prediction frame}

\item{family}{Not used, just for compatibility}

\item{obsWeights}{Not used, just for compatibility}

\item{b}{Interaction coefficient for L01 and L11}

\item{b0}{The intercept}

\item{r}{The exponent of the rate at which error goes to zero as function of n.}

\item{cons}{The constant in front of the rate (default = 1)}
}
\value{
A \code{list} with predictions equal to truth + error
}
\description{
A function to generate an estimate of g0n and g1n that is equal to 
g00 and g10 plus an error term. The function is written for compatibility
with the drinf.tmle function, i.e., as though it were a SuperLearner wrapper.
If \code{newX} is two columns, returns g0n, if four columns, returns g1n.
}
