% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSim.R
\name{truthQ}
\alias{truthQ}
\title{truthQ}
\usage{
truthQ(Y, X, newX, family, obsWeights, ba = -0.25, b = 0.5, r = -1/4,
  cons = 2, ...)
}
\arguments{
\item{Y}{The outcome}

\item{X}{The predictors}

\item{newX}{The prediction frame}

\item{family}{Not used, just for compatibility}

\item{obsWeights}{Not used, just for compatibility}

\item{ba}{The effect of each treatment on the outcome. See \code{makeData}
to see how it is used.}

\item{b}{Interaction coefficient for L01 and L11}

\item{r}{The exponent of the rate at which error goes to zero as function of n.}
}
\value{
A \code{list} with predictions equal to truth + error
}
\description{
A function to generate an estimate of g0n and g1n that is equal to 
Q10 and Q20 plus an error term. The function is written for compatibility
with the drinf.tmle function, i.e., as though it were a SuperLearner wrapper.
If \code{newX} is two columns, returns Q1n, if four columns, returns Q2n.
}

