% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universalStep.R
\name{universalStep}
\alias{universalStep}
\title{universalStep}
\usage{
universalStep(L2, Qn, gn, Qnr.gnr, PnDQ2, PnDQ1, PnDg1, PnDg0, normPnD,
  dx = 0.005, tolg, tolQ, ...)
}
\arguments{
\item{Qn}{A \code{list} of current estimates of Q2n and Q1n}

\item{gn}{A \code{list} of current estimates of g0n and g1n}

\item{Qnr.gnr}{A \code{list} of current estimates of reduced dim. regressions}

\item{normPnD}{A \code{numeric} giving the norm of PnD}

\item{dx}{A \code{numeric} step size}

\item{tolg}{A \code{numeric} indicating the truncation level for conditional treatment probabilities.}

\item{tolQ}{A \code{numeric} indicating the truncation level for transformed outcome regressions.}

\item{PnD}{A \code{matrix} of dimension 3x1 that is the empirical mean of the relevant
portions of the influence function (one for original EIF, one for extra G terms, 
one for extra Q terms).}
}
\description{
Take one step in the practical construction of the universally 
least favorable submodel.
}
