% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnegloglik.R
\name{wnegloglik}
\alias{wnegloglik}
\title{wnegloglik}
\usage{
wnegloglik(epsilon, weight, Y, offset)
}
\arguments{
\item{epsilon}{The scalar parameter of the fluctuation submodel.}

\item{weight}{The \code{vector} of weights, i.e., the clever covariates.}

\item{Y}{The \code{vector} of regression outcomes.}

\item{offset}{The \code{vector} of offsets.}
}
\value{
A \code{numeric} value of the negative log-likelihood loss
}
\description{
A function that computes the weighted negative log-likelihood loss of the 
intercept only fluctuation model.
}
